fix(mypy): batch 3 (50->47) — codemaps/optimization/agent_ui/spectral/sparql/team_tools/owl/approval
.data->.output (codemaps), float() wrap, agent_ui chunk encoding, np.array(dists), sparql form-value
str narrowing, team_tools dynamic setattr, owl_bridge importlib (avoid API-mismatch), approval bool.

+++ agent_utilities/knowledge_graph/memory/optimization_engine.py
@@ -356,7 +356,7 @@ class MemoryOptimizationEngine:
         min_p = 1.0
         for _ in range(self._n_projections):
             direction = rng.standard_normal(d)
-            direction = direction / (np.linalg.norm(direction) or 1.0)
+            direction = float(direction / (np.linalg.norm(direction) or 1.0))
             projection = centered @ direction
             p_value = self._simplified_normality_p(projection)
             min_p = float(min(min_p, p_value))
